automatically set the SPEC file version based on existing setup.py logic
This borrows an approach we use in Image Factory which we, in turn, borrowed
from the Oz tool.  It has worked well.

The special case of release='0' allows frequent RPM rebuilds during development
with sensible version/release progression.  That is, a quick change and rpm rebuild
automatically results in RPM packages that have a "newer" release.

Formal RPM releases can be done by simply setting release to a non-zero value either
in master or in a release branch.